commit edits

+++ qumba/solve.py
@@ -30,8 +30,8 @@ else:
     njit = lambda f : f
 
 
-int_scalar = numpy.int64
-#int_scalar = numpy.int8 # careful with .sum() !
+#int_scalar = numpy.int64
+int_scalar = numpy.int8 # careful with .sum() !
 
 
 def array2(items):
